fix: move ServiceInstanceIdResourceProvider to public resources package and instantiate directly
This fixes the ServiceLoader failure by:
- Moving ServiceInstanceIdResourceProvider back to public resources package as public final class
- Removing it from the SPI service loader registration
- Adding direct instantiation in ResourceConfiguration with conditional application
- This avoids ServiceLoader public no-arg constructor requirement while keeping functionality
- Updated test to disable the provider where needed to preserve original expectations
