+++ sdk-extensions/autoconfigure/src/main/resources/META-INF/services/io.opentelemetry.sdk.autoconfigure.spi.ResourceProvider
@@ -1,2 +1,1 @@
 io.opentelemetry.sdk.autoconfigure.EnvironmentResourceProvider
-io.opentelemetry.sdk.autoconfigure.internal.ServiceInstanceIdResourceProvider
